move page routing to root
