+++ pages/p/preview/_PagePreviewToken.vue
@@ -6,7 +6,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import PageSlug from '~/pages/p/_PageSlug.vue'
+import PageSlug from '~/pages/_PageSlug.vue'
 
 export default Vue.extend({
   name: 'PagePreviewToken',
